Derive src_frame signals from FIFO valid.
Should finally fix off-by-one problems and early termination of input
signals. The problem was that with data going through a 4:1 buffer, the
trigger frame goes low long before the buffer is actually empty. So, for
now we'll just use the buffer/FIFO valid signal as a proxy for frame.

The second problem we are fixing here is that the first-word-fall-thru
FIFO causes the first data point to be read twice. Using the AND of rden
and vld should fix this.

--BRJ and CAR