Add small SRL fifo on input to VitaFramer
Acts as input register and allows backpressure during header to be absorbed.
Also change output to simple register instead of srl register.

The SRL fifo seems to have undefined behaviour in simulation when empty.
I think it is just complaining about the pointer underflowing and it should be fine
in synthesis as it will point to the zeros high shift register entry.